requirements.txt
Django==3.0.5
django-crispy-forms==1.9.0
